almost gradient

+++ fdf.h
@@ -100,7 +100,7 @@ void	init_params(t_fdf *fdf);
 void	get_points(t_map *map, char *file, t_fdf *fdf);
 t_color	hex_to_color(char *n);
 int		rgb_from(char rgb, int color);
-t_step	def_step(t_point *i, t_point *f, int size);
+t_step	def_step(t_point *i, t_point *f, float size);
 int	get_color(t_step *step, t_color color, int current);
 // calculation
 void	calculate_delta(t_fdf *fdf);
